+++ packages/common-ethereum/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Update header year to 2025 (#362)
+- Update `@subql/common` (#363)
 
 ## [4.6.1] - 2025-01-28
 ### Changed